+++ subscriptions_service/src/main/java/com/subOne/subscriptions_service/client/WebClientService.java
@@ -6,7 +6,7 @@ import reactor.core.publisher.Mono;
 import java.util.List;
 
 public interface WebClientService {
-    Mono<Void> checkUserInGroup(Long groupId, Jwt jwt);
-    Mono<Void> checkUserIsOwnerGroup(Long groupId, Jwt jwt);
-    Mono<List<Long>> getGroupsIdByOwnerId(Jwt jwt);
+    Mono<Void> checkUserInGroup(Integer groupId, Jwt jwt);
+    Mono<Void> checkUserIsOwnerGroup(Integer groupId, Jwt jwt);
+    Mono<List<Integer>> getGroupsIdByOwnerId(Jwt jwt);
 }
